+++ doc/src/changelog.md
@@ -3,6 +3,7 @@ Version History
 
 ### Embree 4.1.1
 -   Improved BVH build performance on many core machines for applications that oversubscribe threads.
+-   Added rtcGetGeometryTransformFromScene API function that can get used inside SYCL kernels.
 
 ### Embree 4.1.0
 -   Added support for Intel® Data Center GPU Max Series.